Add use effect
Create a user login status event listener,

After login, checkUser will be called to check
whether firestore has registration information.

If not registered, create a new user,

Then get the user's shopping cart and initialize the cartContext,

Finally, the user avatar changes depending on whether the user is logged in or not.